discord_bot/requirements.txt
interactions.py
pynng
uvloop
python-dotenv